Move src/lib/utils.ts to src/lib/utils/functions.ts
We probably don't want a utils directory and a utils file.
